Split widget actions into basic and standard actions to avoid circular dependency issues
